requirements.txt
streamlit
sentence-transformers
chromadb
PyMuPDF
spacy
groq
python-dotenv
